requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
